refactor: move ShapeDict to _utils.shape to break import cycle
_label_file imported the labelme._utils package, whose __init__ imports
shape.py, which needed ShapeDict back from _label_file. shape.py worked
around the runtime cycle with a TYPE_CHECKING guard and a FIXME.

ShapeDict is the data type shapes_to_label already operates on, so move
its definition down into _utils.shape and have _label_file import it from
there. The dependency is now one-directional and the guard is gone.

Closes #2252
